Remove the `::8` for 8-bit integers in binaries
The default is 8, so this generally isn't specified. This had a side
benefit of reducing the line length in improved formatting in a couple places.

There were a few places where the ::8 looks like it may have been done
for clarity, but I changed all of the places for consistency.