read by type: Fix implementation
This change fixes an issue where `ATT_READ_BY_TYPE_RSP` PDU did not
contain the value of the characteristic selected. Bluetooth
specification "4.8.2 Read Using Characteristic UUID" defines that the
value of the characteristic is added to the PDU if only one
characteristic is selected.

This issue caused empty values in the iOS device list and passkey entry
form as iOS is using this method to read the devices name before
characteristics are discovered.

Signed-off-by: Markus Hutzler <markus@keeplabs.com>